# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ETMS API
  version: "1.0.0"
  description: >
    Employee Transportation Management System REST API. Human-readable contract
    in docs/etms/06-api-spec.md; this file covers the endpoints implemented in the
    backend service. All /v1 routes require a Bearer JWT carrying tenant + permissions.
servers:
  - url: https://api.{region}.etms.app/v1
    variables: { region: { default: dev } }
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
security:
  - bearerAuth: []
paths:
...
  /v1/trips/{id}/assign:
    post:
      summary: Dispatch — assign vehicle + driver (validates capacity & driver state)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Assigned }
        "409": { description: Trip not in an assignable state }
        "422": { description: Vehicle/driver invalid }
  /v1/trips/{id}/start:
    post:
      summary: Start an assigned trip (lifecycle)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Started }, "409": { description: Wrong state } }
  /v1/trips/{id}/complete:
    post:
      summary: Complete an in-progress trip (lifecycle)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Completed }, "409": { description: Wrong state } }
  /v1/drivers/{id}/verify:
    post:
      summary: Verify or reject a driver (sensitive; MFA step-up per policy)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Updated }, "404": { description: Not found } }
...
  /v1/incidents/{id}/resolve:
    post:
      summary: Resolve an incident
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Resolved } }
...